FINAL ILP CW2 VERSION

+++ src/main/java/uk/ac/ed/acp/cw2/dtos/NextPositionDto.java
@@ -13,6 +13,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class NextPositionDto {
-    private @NotNull @Valid PositionDto start;
-    private @NotNull @Min(value=0) @Max(value=360) int angle;
+  private @NotNull @Valid PositionDto start;
+  private @NotNull @Min(value = 0) @Max(value = 360) double angle;
 }
